workflow: refactor JIT to interpreter
Yep, time was wasted (not really, it was fun) on the JIT implementation,
but after carefully looking at the generated code, it looked really
similar to the code that would have been compiled from an interpreter.

An interpreter solution is much simpler to maintain than JIT, so I
decided to refactor!