server: use rename to output the special `@timestamp` field on union
Each subquery in a union needs to rename `@timestamp` into the
corresponding field in that collection (e.g. quickwit index).

But to output to the user `@timestamp` and to unify the timestamp field
from all the union subqueries to the same name, we add a rename workflow
step to transform it back to `@timestamp`.